Moved source, requeted url and links from the mixin  iqm2 to the every spider
